XEN_SYSCTL_numainfo: memsize: Count e820 RAM pages of NUMA nodes
The NUMA node memsize values are returned by XEN_SYSCTL_numainfo
and currently it is calculated from node_spanned_pages, that is
usually wrong for the 1st NUMA node:

XEN_SYSCTL_numainfo returns memfree and memsize of each NUMA node,
and memsize returns NODE_DATA->node_spanned_pages that includes
memory holes which are not usable memory.

These holes cause that the 1st NUMA node is reported to be e.g.
~2GB larger than it really is.

Those holes are part of the e820 table and the SRAT and
nodes_cover_memory which that the SRAT covers all E820_RAM
areas, but E820_RAM may omit more areas (reserved),
so the e820 table is more exact.

This patch counts the present pages in the NUMA nodes based on
the e820 RAM entries, which may report e.g. 63.5 GB for node0
where the SRAT entries my report 64GB, but the 63.5 GB is what
is actually accessible as E820_RAM.