NUMA: Introduce NODE_DATA->node_present_pages(RAM pages)
Xen tracks the total span of physical memory space of each NUMA node,
but the PFN span includes large MMIO holes in e.g. the first NUMA node.
Thus, the span is not the same as the amount of usable RAM of a node.

Xen does not need it, but NUMA node memory amount can be helpful for
management tools and HW information tools like hwloc/lstopo with its
Xen backend for Dom0: https://github.com/xenserver-next/hwloc/

Introduce node_present_pages to node_data[]:
On boot, set the count of usable PFNs and update it on memory_add().

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>
---
Changes in v5:

- Checked node_present_pages on NUMA HW.
- Code and comments are exactly as suggested & affirmed in v4 review.
- Comment adjustments on commit are ok, but keep the code as-is: e.g.:
  'err' needs to be defined outside the loop to be in scope for check.