SYSCTL_numainfo.memsize: Switch from spanned to present memory
This is part of my submission to fix the NUMA node memsize retured
in struct xen_sysctl_meminfo[].memsize by the XEN_SYSCTL_numainfo
hypercall to not count MMIO memory holes etc but only memory pages.

For this, we introduced NODE_DATA->node_present_pages as a rereq.
With the prereq merged in master, I send this 2nd part for review:

This RFC is for changing the value of xen_sysctl_meminfo[]->memsize
from NODE_DATA->node_spanned_pages << PAGE_SHIFT
  to NODE_DATA->node_present_pages << PAGE_SHIFT
for returing present pages instead of spanned pages.

This is how the contensts of struct xen_sysctl_meminfo[] look in xl:

xl info -n:
[...]
node:    memsize    memfree    distances
   0:  -> 67584 <-   60672      10,21
   1:     65536      60958      21,10

The -> memsize <- marked here is the value that we'd like to fix:

The current value based node_spanned_pages is often 2TB too large.

We're currently not using these memsize values in XenServer according
to my own code review and Andrew Cooper.

I would think that the same is likely true for other Xen toolstacks,
but of course to review this change is the purpose of this RFC.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>