From a3718753b6ad20381effcac6c8d119b7f76af22d Mon Sep 17 00:00:00 2001 [PATCH] x86/hvmloader: account for external components consuming PCI
 window space
MIME-Version: 1.0
Content-Type: text/plain; charset=UTF-8
Content-Transfer-Encoding: 8bit

It's possible for external components to steal part of the end of the low
PCI MMIO window, they would account for this by setting the stolen region
as revered in the e820 memory map.

Add some logic in hvmloader to check whether the chosen PCI window end
address overlaps with reserved regions, and it that case adjust the end of
the window to take the reservation into account.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>