From 59cf7857d08cf111ce0575fff7d2c11351607358 Mon Sep 17 00:00:00 2001 [PATCH v2 4/7] xen/page_alloc: Pass node to adjust_tot_pages,
 set_outstanding_pages

domain_adjust_tot_pages() was used to adjust the total pages of a domain
and when adding pages for updating outstanding claims, the NUMA node
needs to be passed. For decreasing the claims, a NUMA node is not used.

To avoid passing NUMA_NO_NODE to each call that reduces the tot_pages,
switch to static inline domain_decrease_total_pages() for decreasing
tot_pages. This improves the legibility of the calls and avoids raising
eyebrows for lots of calls where NUMA_NO_NODE would be passed to a call
to update the tot_pages. With the comment added to the static inline
the situation is explained for all reviewers and and the legibility
of the code is improved as a bonus.

The old function name domain_adjust_tot_pages() is removed and the new
function is renamed to domain_update_total_pages(). It adds the node
argument and is only used for call sites that increase the total pages,
which need to pass the NUMA node of the pages added to the domain.

For consistency, also add domain_increase_total_pages() as a static inline
for calls that increase the tot_pages, which also need to pass the NUMA node.

domain_set_outstanding_pages() sets the amount of outstanding claims
of a domain. We pass the node on which to to stake an claim,
or NUMA_NO_NODE for host-wide claims.

No functional change, as neither function uses the arguments
for anything yet. It's a prerequisite to simplify for the next
patch that introduces per-node claim counts.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>