From cc44cbe032e6dfea7760558b757414885a5bb2d3 Mon Sep 17 00:00:00 2001 [PATCH v2 7/7] xen: New API to claim memory for a domain using
 XEN_DOMCTL_claim_memory

Add the new hypercall requested during the review of the v1 series
do not require changing the API for multi-node claims.

The hypercall receives a number of claims, intented to be one claim per
NUMA node, and limited to one claim for now. The changes to update the
NUMA claims management to handle updating the claims for multiple
NUMA nodes of a domain at once are deferred to the next series.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>
Cc: Alejandro Vallejo <alejandro.garciavallejo@amd.com>