From a3d5e5effb933b865f8545f5354f3a8f8606e971 Mon Sep 17 00:00:00 2001 [PATCH v2 1/7] xen/page_alloc: Simplify domain_adjust_tot_pages for
 future changes

domain_adjust_tot_pages() is used to update the domain's total pages
after allocating and freeing memory.

Simplify the design for updating it for single and even more so
for multi-node claims regarding the case where we could have
allocated more memory than we had claims left.

Replace it with min_t() to avoid reducing the outstadings claims
by more than we had left to claim:

When domain memory is freed, we skip changing the claim. Thus, this
only handles reducing the claims after allocating. So, min_t() is fine.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>